GO-1240: remove status watcher
