+++ core/syncstatus/service.go
@@ -2,24 +2,22 @@ package syncstatus
 
 import (
 	"context"
-	"sync"
+	"fmt"
 
 	"github.com/anyproto/any-sync/app"
-	"github.com/anyproto/any-sync/commonspace"
 
 	"github.com/anyproto/anytype-heart/core/block/cache"
+	"github.com/anyproto/anytype-heart/core/block/editor/smartblock"
+	"github.com/anyproto/anytype-heart/core/domain"
 	"github.com/anyproto/anytype-heart/core/filestorage/filesync"
-	"github.com/anyproto/anytype-heart/core/syncstatus/objectsyncstatus"
-	"github.com/anyproto/anytype-heart/pkg/lib/localstore/objectstore"
+	"github.com/anyproto/anytype-heart/core/syncstatus/filesyncstatus"
+	"github.com/anyproto/anytype-heart/pkg/lib/bundle"
+	"github.com/anyproto/anytype-heart/util/pbtypes"
 )
 
 const CName = "status"
 
 type Service interface {
-	Watch(spaceId string, id string, filesGetter func() []string) (new bool, err error)
-	Unwatch(spaceID string, id string)
-	RegisterSpace(space commonspace.Space, sw objectsyncstatus.StatusWatcher)
-
 	app.ComponentRunnable
 }
 
@@ -27,23 +25,15 @@ var _ Service = (*service)(nil)
 
 type service struct {
 	fileSyncService filesync.FileSync
-
-	objectWatchersLock sync.Mutex
-	objectWatchers     map[string]objectsyncstatus.StatusWatcher
-
-	objectStore  objectstore.ObjectStore
-	objectGetter cache.ObjectGetter
+	objectGetter    cache.ObjectGetter
 }
 
 func New() Service {
-	return &service{
-		objectWatchers: map[string]objectsyncstatus.StatusWatcher{},
-	}
+	return &service{}
 }
 
 func (s *service) Init(a *app.App) (err error) {
 	s.fileSyncService = app.MustComponent[filesync.FileSync](a)
-	s.objectStore = app.MustComponent[objectstore.ObjectStore](a)
 	s.objectGetter = app.MustComponent[cache.ObjectGetter](a)
 	s.fileSyncService.OnUploaded(s.onFileUploaded)
 	s.fileSyncService.OnUploadStarted(s.onFileUploadStarted)
@@ -59,47 +49,35 @@ func (s *service) Name() string {
 	return CName
 }
 
-func (s *service) RegisterSpace(space commonspace.Space, sw objectsyncstatus.StatusWatcher) {
-	s.objectWatchersLock.Lock()
-	defer s.objectWatchersLock.Unlock()
-	s.objectWatchers[space.Id()] = sw
-}
-
-func (s *service) UnregisterSpace(space commonspace.Space) {
-	s.objectWatchersLock.Lock()
-	defer s.objectWatchersLock.Unlock()
-
-	// TODO: [MR] now we can't set a nil update receiver, but maybe it doesn't matter that much
-	//  and we can just leave as it is, because no events will come through
-	delete(s.objectWatchers, space.Id())
-}
-
-func (s *service) Unwatch(spaceID string, id string) {
-	s.unwatch(spaceID, id)
-}
-
-func (s *service) Watch(spaceId string, id string, filesGetter func() []string) (new bool, err error) {
-	s.objectWatchersLock.Lock()
-	defer s.objectWatchersLock.Unlock()
-	objectWatcher := s.objectWatchers[spaceId]
-	if objectWatcher != nil {
-		if err = objectWatcher.Watch(id); err != nil {
-			return false, err
-		}
-	}
-	return true, nil
-
-}
-
-func (s *service) unwatch(spaceID string, id string) {
-	s.objectWatchersLock.Lock()
-	defer s.objectWatchersLock.Unlock()
-	objectWatcher := s.objectWatchers[spaceID]
-	if objectWatcher != nil {
-		objectWatcher.Unwatch(id)
-	}
-}
-
 func (s *service) Close(ctx context.Context) (err error) {
+	return nil
+}
+
+func (s *service) onFileUploadStarted(objectId string, _ domain.FullFileId) error {
+	return s.indexFileSyncStatus(objectId, filesyncstatus.Syncing)
+}
+
+func (s *service) onFileUploaded(objectId string, _ domain.FullFileId) error {
+	return s.indexFileSyncStatus(objectId, filesyncstatus.Synced)
+}
+
+func (s *service) onFileLimited(objectId string, _ domain.FullFileId, bytesLeftPercentage float64) error {
+	return s.indexFileSyncStatus(objectId, filesyncstatus.Limited)
+}
+
+func (s *service) indexFileSyncStatus(fileObjectId string, status filesyncstatus.Status) error {
+	err := cache.Do(s.objectGetter, fileObjectId, func(sb smartblock.SmartBlock) (err error) {
+		prevStatus := pbtypes.GetInt64(sb.Details(), bundle.RelationKeyFileBackupStatus.String())
+		newStatus := int64(status)
+		if prevStatus == newStatus {
+			return nil
+		}
+		st := sb.NewState()
+		st.SetDetailAndBundledRelation(bundle.RelationKeyFileBackupStatus, pbtypes.Int64(newStatus))
+		return sb.Apply(st)
+	})
+	if err != nil {
+		return fmt.Errorf("get object: %w", err)
+	}
 	return nil
 }
