+++ space/spacecore/service.go
@@ -81,12 +81,6 @@ type service struct {
 	peerService          peerservice.PeerService
 	poolManager          PoolManager
 	streamHandler        *streamHandler
-	syncStatusService    syncStatusService
-}
-
-type syncStatusService interface {
-	RegisterSpace(space commonspace.Space, sw objectsyncstatus.StatusWatcher)
-	UnregisterSpace(space commonspace.Space)
 }
 
 func (s *service) Init(a *app.App) (err error) {
@@ -101,7 +95,6 @@ func (s *service) Init(a *app.App) (err error) {
 	s.spaceStorageProvider = a.MustComponent(spacestorage.CName).(storage.ClientStorage)
 	s.peerStore = a.MustComponent(peerstore.CName).(peerstore.PeerStore)
 	s.peerService = a.MustComponent(peerservice.CName).(peerservice.PeerService)
-	s.syncStatusService = app.MustComponent[syncStatusService](a)
 	localDiscovery := a.MustComponent(localdiscovery.CName).(localdiscovery.LocalDiscovery)
 	localDiscovery.SetNotifier(s)
 	s.streamHandler = &streamHandler{spaceCore: s}
@@ -242,7 +235,7 @@ func (s *service) loadSpace(ctx context.Context, id string) (value ocache.Object
 	if err != nil {
 		return
 	}
-	ns, err := newAnySpace(cc, s.syncStatusService, statusService)
+	ns, err := newAnySpace(cc)
 	if err != nil {
 		return
 	}
