+++ space/spacecore/space.go
@@ -9,17 +9,12 @@ import (
 	"github.com/anyproto/anytype-heart/core/syncstatus/objectsyncstatus"
 )
 
-func newAnySpace(cc commonspace.Space, status syncStatusService, statusWatcher objectsyncstatus.StatusWatcher) (*AnySpace, error) {
-	return &AnySpace{
-		Space:         cc,
-		status:        status,
-		statusWatcher: statusWatcher,
-	}, nil
+func newAnySpace(cc commonspace.Space) (*AnySpace, error) {
+	return &AnySpace{Space: cc}, nil
 }
 
 type AnySpace struct {
 	commonspace.Space
-	status        syncStatusService
 	statusWatcher objectsyncstatus.StatusWatcher
 }
 
@@ -28,7 +23,6 @@ func (s *AnySpace) Init(ctx context.Context) (err error) {
 	if err != nil {
 		return
 	}
-	s.status.RegisterSpace(s, s.statusWatcher)
 	return
 }
 
@@ -37,6 +31,5 @@ func (s *AnySpace) TryClose(objectTTL time.Duration) (close bool, err error) {
 }
 
 func (s *AnySpace) Close() (err error) {
-	s.status.UnregisterSpace(s)
 	return s.Space.Close()
 }
